FPGA: free up the USP_SPARE0,1 lines on CW610.
FPGA reset is now done via a register write.
ALE isn't really needed after all. The only scenario impacted by this
change is if you wanted to repeatedly read or write the first byte of a
multi-byte register; you can't do that anymore. Now, the byte offset
gets reset to 0 only when a new address (different from the previous
one) is seen.
The objective of these changes is that the USB_SPARE* lines will be
available to enable a new faster FIFO read mode.