Verilog: add missing sync frame combination.
Could cause recording to be enabled during a string of sync frames in a
way that resulted in partial sync frames being recorded which could then
lead to problems in parsing the trace data. Unlikely to happen in real
life(?) because it involves a mix of short and long sync frames which I
haven't seen.